backend/requirements.txt
# Renta-Car Backend Dependencies
# For Azure App Service deployment

Flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.1.0
pyodbc==5.2.0
Werkzeug==3.1.3
PyJWT==2.10.1
reportlab==4.3.1
Pillow==11.2.1
gunicorn==23.0.0
